+++ apps/exams/serializers/__init__.py
@@ -1,16 +1,16 @@
-from .admin_submission_serializers import AdminExamSubmissionListResponseSerializer
+from .admin.submissions import AdminExamSubmissionListResponseSerializer
 from .error_serializers import ErrorDetailSerializer, ErrorResponseSerializer
-from .exam_status import ExamStatusResponseSerializer
 from .exam_submission import (
     ExamSubmissionCreateResponseSerializer,
     ExamSubmissionCreateSerializer,
 )
-from .submit_exam import (
+from .student.exam_status import ExamStatusResponseSerializer
+from .student.submit_exam import (
     AnswerItemSerializer,
     SubmitExamRequestSerializer,
     SubmitExamResponseSerializer,
 )
-from .take_exam import (
+from .student.take_exam import (
     CheckCodeRequestSerializer,
     TakeExamRequestSerializer,
     TakeExamResponseSerializer,
