+++ apps/exams/services/__init__.py
@@ -1,5 +1,5 @@
-from .submit_exam import submit_exam
-from .take_exam import build_take_exam_response, take_exam
+from .student.submit_exam import submit_exam
+from .student.take_exam import build_take_exam_response, take_exam
 
 __all__ = [
     "build_take_exam_response",
