+++ apps/exams/services/admin/exams_delete.py
