+++ apps/exams/tests/admin/test_deployments_status.py
