+++ apps/exams/tests/admin/test_exams_delete.py
