+++ apps/exams/tests/admin/test_questions_delete.py
