+++ apps/exams/urls/admin.py
@@ -1,19 +1,19 @@
 from django.urls import path
 
-from apps.exams.views.admin_exam_delete_views import AdminExamDeleteAPIView
+from apps.exams.views.admin.exams_delete import AdminExamDeleteAPIView
+from apps.exams.views.admin.deployments_status import (
+    AdminExamDeploymentStatusAPIView,
+)
+from apps.exams.views.admin.deployments import (
+    AdminExamDeploymentCreateAPIView,
+)
+from apps.exams.views.admin.exams import AdminExamCreateAPIView
+from apps.exams.views.admin.questions_delete import AdminExamQuestionDeleteAPIView
+from apps.exams.views.admin.questions import AdminExamQuestionCreateAPIView
+from apps.exams.views.admin.submissions import AdminExamSubmissionListAPIView
 from apps.exams.views.admin_exam_deployment_detail_views import (
     AdminExamDeploymentDetailAPIView,
 )
-from apps.exams.views.admin_exam_deployment_status_views import (
-    AdminExamDeploymentStatusAPIView,
-)
-from apps.exams.views.admin_exam_deployment_views import (
-    AdminExamDeploymentCreateAPIView,
-)
-from apps.exams.views.admin_exam_views import AdminExamCreateAPIView
-from apps.exams.views.admin_question_delete_views import AdminExamQuestionDeleteAPIView
-from apps.exams.views.admin_question_views import AdminExamQuestionCreateAPIView
-from apps.exams.views.admin_submission_views import AdminExamSubmissionListAPIView
 
 urlpatterns = [
     path("exams/<int:exam_id>/questions/", AdminExamQuestionCreateAPIView.as_view(), name="admin-exam-question-create"),
