+++ apps/exams/urls/student.py
@@ -1,10 +1,10 @@
 from django.urls import path
 
 from apps.exams.views import CheckCodeAPIView, ExamListView, TakeExamAPIView
-from apps.exams.views.cheating_views import ExamCheatingUpdateAPIView
 from apps.exams.views.exam_result import ExamSubmissionDetailView
 from apps.exams.views.exam_submission import ExamSubmissionCreateAPIView
-from apps.exams.views.status_views import ExamStatusCheckAPIView
+from apps.exams.views.student.cheating import ExamCheatingUpdateAPIView
+from apps.exams.views.student.status import ExamStatusCheckAPIView
 
 app_name = "exams"
