+++ apps/exams/views/__init__.py
@@ -1,9 +1,9 @@
-from .check_code import CheckCodeAPIView
-from .exam_list import ExamListView
-from .exam_status import ExamStatusAPIView
 from .exam_submission import ExamSubmissionCreateAPIView
-from .submit_exam import SubmitExamAPIView
-from .take_exam import TakeExamAPIView
+from .student.check_code import CheckCodeAPIView
+from .student.deployments import ExamListView
+from .student.exam_status import ExamStatusAPIView
+from .student.submit_exam import SubmitExamAPIView
+from .student.take_exam import TakeExamAPIView
 
 __all__ = [
     "CheckCodeAPIView",
