+++ apps/exams/views/admin/deployments.py
@@ -6,12 +6,12 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from apps.exams.permissions import IsExamStaff
-from apps.exams.serializers.admin_exam_deployment_serializers import (
+from apps.exams.serializers.admin.deployments import (
     AdminExamDeploymentCreateRequestSerializer,
     AdminExamDeploymentCreateResponseSerializer,
 )
 from apps.exams.serializers.error_serializers import ErrorResponseSerializer
-from apps.exams.services.admin_exam_deployment_service import (
+from apps.exams.services.admin.deployments import (
     ExamDeploymentConflictError,
     ExamDeploymentNotFoundError,
     create_exam_deployment,
