+++ apps/exams/views/admin/exams.py
@@ -9,12 +9,12 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from apps.exams.permissions import IsExamStaff
-from apps.exams.serializers.admin_exam_serializers import (
+from apps.exams.serializers.admin.exams import (
     AdminExamCreateRequestSerializer,
     AdminExamCreateResponseSerializer,
 )
 from apps.exams.serializers.error import ErrorDetailSerializer
-from apps.exams.services.admin_exam_service import (
+from apps.exams.services.admin.exams import (
     ExamCreateConflictError,
     ExamCreateNotFoundError,
     create_exam,
