+++ apps/exams/views/student/cheating.py
@@ -11,7 +11,7 @@ from rest_framework.views import APIView
 from apps.exams.constants import ExamStatus
 from apps.exams.models import ExamDeployment, ExamSubmission
 from apps.exams.permissions import IsStudentRole
-from apps.exams.serializers.cheating_serializers import (
+from apps.exams.serializers.student.cheating import (
     ExamCheatingRequestSerializer,
     ExamCheatingResponseSerializer,
 )
@@ -19,7 +19,7 @@ from apps.exams.serializers.error_serializers import (
     ErrorDetailSerializer,
     ErrorResponseSerializer,
 )
-from apps.exams.services.exam_status_service import is_exam_active
+from apps.exams.services.student.status import is_exam_active
 from apps.exams.services.grading import grade_submission
